Apply proxy mirror on profile switch; keep BrowserPanel under length budget
switchToProfile swaps the panel onto the selected profile's
WKWebsiteDataStore but never re-applied the proxy configuration, so a
fresh profile store had no explicit proxyConfigurations and the loopback
bypass was lost right after switching profiles (caught by Codex
autoreview). Proxy configurations live on the data store, so every path
that assigns a different store — init, workspace reattach, and now
profile switch — funnels through applyProxyConfigurationIfAvailable();
webview rebuilds that keep the same store (crash recovery, workspace
context reset) inherit the store's existing configuration.

Also brings BrowserPanel.swift back under the Swift file-length budget
(workflow-guard-tests) by shrinking instead of bumping the budget: the
live-settings reader moves into BrowserSystemProxyMirror as
currentProxyConfigurations(), and the BrowserProxyEndpoint /
BrowserRemoteWorkspaceStatus value types move into their own files per
the one-type-per-file rule.

Issue: https://github.com/manaflow-ai/cmux/issues/5888
