+++ Sources/Panels/BrowserPanel.swift
@@ -55,18 +55,6 @@ private struct BrowserFocusModePlainEscapeEventFingerprint: Equatable {
     }
 }
 
-struct BrowserProxyEndpoint: Equatable {
-    let host: String
-    let port: Int
-}
-
-struct BrowserRemoteWorkspaceStatus: Equatable {
-    let target: String
-    let connectionState: WorkspaceRemoteConnectionState
-    let heartbeatCount: Int
-    let lastHeartbeatAt: Date?
-}
-
 enum GhosttyBackgroundTheme {
     static func clampedOpacity(_ opacity: Double) -> CGFloat {
         WindowAppearanceSnapshot.clampedOpacity(opacity)
@@ -4764,19 +4752,10 @@ final class BrowserPanel: Panel, ObservableObject {
 
         let store = webView.configuration.websiteDataStore
         guard let endpoint = remoteProxyEndpoint else {
-            // Local workspaces: mirror an active macOS system proxy with
-            // loopback excluded so http://localhost dev servers stay
-            // reachable under a global proxy (issue #5888). With no mappable
-            // system proxy this stays empty and WebKit keeps following the
-            // system proxy as before.
-            let configurations = Self.localWorkspaceProxyConfigurations()
-            store.proxyConfigurations = configurations
-#if DEBUG
-            cmuxDebugLog(
-                "browser.systemProxyMirror.apply panel=\(id.uuidString.prefix(5)) " +
-                "mirrored=\(configurations.isEmpty ? 0 : 1)"
-            )
-#endif
+            // Mirror an active system proxy with loopback excluded so local
+            // dev servers stay reachable under a global proxy (issue #5888);
+            // empty (system proxy as before) when no faithful mirror exists.
+            store.proxyConfigurations = BrowserSystemProxyMirror.currentProxyConfigurations()
             return
         }
 
@@ -4792,23 +4771,6 @@ final class BrowserPanel: Panel, ObservableObject {
         let socks = ProxyConfiguration(socksv5Proxy: nwEndpoint)
         let connect = ProxyConfiguration(httpCONNECTProxy: nwEndpoint)
         store.proxyConfigurations = [socks, connect]
-    }
-
-    /// Builds the explicit proxy configurations for a local-workspace data
-    /// store by mirroring the active macOS system proxy, excluding loopback
-    /// and the user's proxy bypass list
-    /// (https://github.com/manaflow-ai/cmux/issues/5888). Returns an empty
-    /// array when no mappable system proxy is active so WebKit falls back to
-    /// the system proxy unchanged.
-    private static func localWorkspaceProxyConfigurations() -> [ProxyConfiguration] {
-        guard let rawSettings = CFNetworkCopySystemProxySettings()?.takeRetainedValue() else {
-            return []
-        }
-        guard let settings = rawSettings as NSDictionary as? [String: Any],
-              let mirror = BrowserSystemProxyMirror(systemProxySettings: settings) else {
-            return []
-        }
-        return mirror.proxyConfigurations()
     }
 
     private func beginDownloadActivity() {
@@ -4938,6 +4900,7 @@ final class BrowserPanel: Panel, ObservableObject {
         refreshWebViewLifecycleState()
 
         bindWebView(replacement)
+        applyProxyConfigurationIfAvailable()
         applyBrowserThemeModeIfNeeded()
 
         if !history.backHistoryURLStrings.isEmpty || !history.forwardHistoryURLStrings.isEmpty {
@@ -5331,9 +5294,7 @@ final class BrowserPanel: Panel, ObservableObject {
         // Keep the local-workspace system-proxy mirror fresh when the user
         // toggles a global proxy or switches network locations mid-session.
         NotificationCenter.default.publisher(for: .browserSystemProxySettingsDidChange)
-            .sink { [weak self] _ in
-                self?.applyProxyConfigurationIfAvailable()
-            }
+            .sink { [weak self] _ in self?.applyProxyConfigurationIfAvailable() }
             .store(in: &webViewCancellables)
 
         // Apply the configured background for the freshly bound webview (covers
