+++ Sources/Panels/BrowserRemoteWorkspaceStatus.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+/// Connection snapshot for the remote workspace a browser panel belongs to.
+struct BrowserRemoteWorkspaceStatus: Equatable {
+    let target: String
+    let connectionState: WorkspaceRemoteConnectionState
+    let heartbeatCount: Int
+    let lastHeartbeatAt: Date?
+}
